Revert "Final version"
