+++ README.md
@@ -66,7 +66,6 @@
 
 - 过程名称：`@localhost/cn.fmsoft.hybridos.hibus/busybox/chdir`
 - 参数：
-   + `euid`：调用者的有效用户ID；
    + `dictName`：要设置的工作目录的全路径；
 ```json
     {
@@ -88,9 +87,7 @@
 
 - 过程名称：`@localhost/cn.fmsoft.hybridos.hibus/busybox/ls`
 - 参数：
-   + `euid`：调用者的有效用户ID；
    + `path`：指定的目录，如果不是以`/`开头，则认为是相对路径；
-   + `option`：`ls`命令参数，目前支持`a`；
 ```json
     {
         "euid": euid,
@@ -150,7 +147,6 @@
 
 - 过程名称：`@localhost/cn.fmsoft.hybridos.hibus/busybox/rm`
 - 参数：
-   + `euid`：调用者的有效用户ID；
    + `fileName`：要删除文件或目录的路径，如果不是以`/`开头，则认为是相对路径；
 ```json
     {
@@ -172,7 +168,6 @@
 
 - 过程名称：`@localhost/cn.fmsoft.hybridos.hibus/busybox/rmdir`
 - 参数：
-   + `euid`：调用者的有效用户ID；
    + `fileName`：要删除空目录的路径，如果不是以`/`开头，则认为是相对路径；
 ```json
     {
@@ -194,7 +189,6 @@
 
 - 过程名称：`@localhost/cn.fmsoft.hybridos.hibus/busybox/mkdir`
 - 参数：
-   + `euid`：调用者的有效用户ID；
    + `fileName`：要创建目录的路径，如果不是以`/`开头，则认为是相对路径；
 ```json
     {
@@ -216,7 +210,6 @@
 
 - 过程名称：`@localhost/cn.fmsoft.hybridos.hibus/busybox/unlink`
 - 参数：
-   + `euid`：调用者的有效用户ID；
    + `fileName`：要删除文件的路径，如果不是以`/`开头，则认为是相对路径；
 ```json
     {
@@ -237,108 +230,11 @@
 #### 修改文件属性
 
 - 过程名称：`@localhost/cn.fmsoft.hybridos.hibus/busybox/touch`
-   + `euid`：调用者的有效用户ID；
    + `fileName`：要修改属性的文件路径，如果不是以`/`开头，则认为是相对路径；
 ```json
     {
         "euid": euid,
         "fileName":"path"
-    }
-```
-- 返回值：
-   + `errCode`：返回错误编码，见附表；
-   + `errMsg`：错误信息。
-```json
-    {
-        "errCode":0,
-        "errMsg":"OK"
-    }
-```
-
-#### 获取当前文件系统的磁盘空间
-
-- 过程名称：`@localhost/cn.fmsoft.hybridos.hibus/busybox/df`
-   + `euid`：调用者的有效用户ID；
-   + `fileName`：要修改属性的文件路径，如果不是以`/`开头，则认为是相对路径；
-```json
-    {
-        "euid": euid,
-        "fileName":"path"
-    }
-```
-- 返回值：
-   + `blocksize`: 磁盘块大小；
-   + `totalsize`: 分区占用块数；
-   + `freeDisk`: 空闲空间；
-   + `availableDisk`: 可使用的空间；
-   + `errCode`：返回错误编码，见附表；
-   + `errMsg`：错误信息。
-```json
-    {
-        "blocksize": 512
-        "totalsize": 123456789
-        "freeDisk": 2333333
-        "availableDisk": 2444444
-        "errCode":0,
-        "errMsg":"OK"
-    }
-```
-
-#### 获取环境变量
-
-- 过程名称：`@localhost/cn.fmsoft.hybridos.hibus/busybox/getenv`
-   + `euid`：调用者的有效用户ID；
-   + `envName`：要获取环境变量的`key`；
-```json
-    {
-        "euid": euid,
-        "envName":"key"
-    }
-```
-- 返回值：
-   + `value`: 环境变量；
-   + `errCode`：返回错误编码，见附表；
-   + `errMsg`：错误信息。
-```json
-    {
-        "value": "value"
-        "errCode":0,
-        "errMsg":"OK"
-    }
-```
-
-#### 设置环境变量
-
-- 过程名称：`@localhost/cn.fmsoft.hybridos.hibus/busybox/setenv`
-   + `euid`：调用者的有效用户ID；
-   + `envName`：要设置环境变量的`key`；
-   + `envValue`：要设置环境变量的`value`；
-```json
-    {
-        "euid": euid,
-        "envName":"key"
-        "envValue":"value"
-    }
-```
-- 返回值：
-   + `errCode`：返回错误编码，见附表；
-   + `errMsg`：错误信息。
-```json
-    {
-        "errCode":0,
-        "errMsg":"OK"
-    }
-```
-
-#### 删除环境变量
-
-- 过程名称：`@localhost/cn.fmsoft.hybridos.hibus/busybox/unsetenv`
-   + `euid`：调用者的有效用户ID；
-   + `envName`：要删除的环境变量的`key`；
-```json
-    {
-        "euid": euid,
-        "envName":"key"
     }
 ```
 - 返回值：
@@ -385,11 +281,10 @@
 | ERROR_BUSYBOX_EFBIG             | -27     | File is too large.                            |
 | ERROR_BUSYBOX_EINTR             | -28     | Interrupted system call.                      |
 | ERROR_BUSYBOX_EMFILE            | -29     | Too many open files.                          |
-| ERROR_BUSYBOX_ENOSYS            | -30     | System does not support.                      |
-| ERROR_BUSYBOX_INVALID_PARAM     | -31     | Invalid parameter.                            |
-| ERROR_BUSYBOX_INSUFFICIENT_MEM  | -32     | Insufficient memory.                          |
-| ERROR_BUSYBOX_CONNECT_HIBUS     | -33     | Can not connect to Hibus server.              |
-| ERROR_BUSYBOX_WRONG_PROCEDURE   | -34     | Wrong procedure name.                         |
-| ERROR_BUSYBOX_WRONG_JSON        | -35     | Wrong Json format.                            |
-| ERROR_BUSYBOX_WORKING_DIRECTORY | -36     | Current working directory has not benn set.   |
-| ERROR_BUSYBOX_UNKONOWN          | -37     | Unknow error.                                 |
+| ERROR_BUSYBOX_INVALID_PARAM     | -30     | Invalid parameter.                            |
+| ERROR_BUSYBOX_INSUFFICIENT_MEM  | -31     | Insufficient memory.                          |
+| ERROR_BUSYBOX_CONNECT_HIBUS     | -32     | Can not connect to Hibus server.              |
+| ERROR_BUSYBOX_WRONG_PROCEDURE   | -33     | Wrong procedure name.                         |
+| ERROR_BUSYBOX_WRONG_JSON        | -34     | Wrong Json format.                            |
+| ERROR_BUSYBOX_WORKING_DIRECTORY | -35     | Current working directory has not benn set.   |
+| ERROR_BUSYBOX_UNKONOWN          | -36     | Unknow error.                                 |
