+++ src/filesystem.h
@@ -1,13 +1,23 @@
 #ifndef __HIBUS_BUSYBOX_FS__
 #define __HIBUS_BUSYBOX_FS__
 
+struct wildcard_list
+{
+    char *wildcard;
+    struct wildcard_list *next;
+};
+
+
+#define	PROCEDURE_PARAM	hibus_conn* conn, const char* from_endpoint, \
+            const char* to_method, const char* method_param, int *err_code
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
 /****************************************************************************
  * Subroutine:  fs_register
- * Function:    register file system procedure and event for busybox runner
+ * Function:    register procedure and event for busybox runner
  * Input:       hibus_context[I], Hibus context
  * Output:      None
  * Description:
@@ -19,7 +29,7 @@ void fs_register(hibus_conn *hibus_context);
 
 /****************************************************************************
  * Subroutine:  fs_revoke
- * Function:    revoke file system procedure and event for busybox runner
+ * Function:    revoke procedure and event for busybox runner
  * Input:       hibus_context[I], Hibus context
  * Output:      None
  * Description:
@@ -27,6 +37,18 @@ void fs_register(hibus_conn *hibus_context);
  * ModifyRecord:
  * *************************************************************************/
 void fs_revoke(hibus_conn *hibus_context);
+
+
+/****************************************************************************
+ * Subroutine:  changeDirectory
+ * Function:    change current working directory.
+ * Input:       Hibus register procedure input
+ * Output:      Hibus register procedure output
+ * Description:
+ * Date:        2022/07/20
+ * ModifyRecord:
+ * *************************************************************************/
+char * changeDirectory(PROCEDURE_PARAM);
 
 
 /****************************************************************************
