+++ src/helper.h
@@ -42,43 +42,6 @@ void change_euid(const char *from_endpoint, uid_t euid);
  * *************************************************************************/
 void restore_euid(void);
 
-
-/****************************************************************************
- * Subroutine:  wildcard_cmp
- * Function:    Judge whether a string matches a wildcard
- * Input:       text[I], the string to be matched
- *              pattern[I], the wildcard
- * Output:      true for matched, otherwise false
- * Description:
- * Date:        2022/07/20
- * ModifyRecord:
- * *************************************************************************/
-bool wildcard_cmp(const char *text, const char *pattern);
-
-
-/****************************************************************************
- * Subroutine:  get_error_code
- * Function:    get error code according to errno
- * Input:       none
- * Output:      error code 
- * Description:
- * Date:        2022/07/20
- * ModifyRecord:
- * *************************************************************************/
-int get_error_code(void);
-
-
-struct env_param * get_env(struct env_param *head, const char *key);
-
-struct env_param *
-add_env(struct env_param **head, const char *key, const char *value);
-
-bool remove_env(struct env_param **head, const char *key);
-
-struct env_param *
-modify_env(struct env_param **head, const char *key, const char *value);
-
-
 #ifdef __cplusplus
 }
 #endif
